+++ hashstash/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.0.0'
+__version__ = '1.0.1'
 from .constants import *
 from .config import *
 from .hashstash import *